gaia: gaia: An R package to estimate crop yield responses to temperature and precipitation

An empirical model to estimate annual crop yield responses to temperature and precipitation.

See also

Author

Maintainer: Mengqi Zhao mengqi.zhao@pnnl.gov (ORCID)

Authors:

Stephanie Waldhoff Stephanie.Waldhoff@pnnl.gov (ORCID)

Claudia Tebaldi claudia.tebaldi@pnnl.gov (ORCID)

Abigail Snyder abigail.snyder@pnnl.gov (ORCID)