weighted_climate

Process standard NetCDF files from ISIMIP to country level climate based on crop harvested area from MIRCA2000

weighted_climate(
  pr_ncdf = NULL,
  tas_ncdf = NULL,
  timestep = "monthly",
  climate_model = "gcm",
  climate_scenario = "rcp",
  time_periods = NULL,
  output_dir = file.path(getwd(), "output"),
  name_append = NULL
)

Arguments

pr_ncdf

Default = NULL. list of paths for precipitation NetCDF files from ISIMIP

tas_ncdf

Default = NULL. list of paths for temperature NetCDF files from ISIMIP

timestep

Default = 'monthly'. string for input climate data time step (e.g., 'monthly', 'daily')

climate_model

Default = NULL. string for climate model (e.g., 'CanESM5')

climate_scenario

Default = NULL. string for climate scenario (e.g., 'ssp245')

time_periods

Default = NULL. vector for years to subset from the climate data. If NULL, use the default climate data period

output_dir

Default = file.path(getwd(), 'output'). String for output directory

name_append

Default = NULL. string for name append to the output folder